% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\name{as.list,feature_data-method}
\alias{as.list,feature_data-method}
\title{Convert feature_data to a list}
\usage{
\S4method{as.list}{feature_data}(x)
}
\arguments{
\item{x}{A `feature_data` object.}
}
\value{
A `list` object.
}
\description{
Coerces a `feature_data` object to a `list` object.
}
