% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\name{as.list,preproc_data-method}
\alias{as.list,preproc_data-method}
\title{Convert preproc_data to a list}
\usage{
\S4method{as.list}{preproc_data}(x)
}
\arguments{
\item{x}{A `preproc_data` object.}
}
\value{
A `list` object.
}
\description{
Coerces a `preproc_data` object to a `list` object.
}
