% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAs.R
\name{as.list}
\alias{as.list}
\alias{as.list,preproc_data-method}
\title{Convert preproc_data to preproc_data_list}
\usage{
\S4method{as.list}{preproc_data}(x)
}
\arguments{
\item{from}{A `preproc_data` object.}
}
\value{
A `preproc_data_list` object.

A `list` object.
}
\description{
Coerces a `preproc_data` object to a `preproc_data_list` object.

Coerces a `preproc_data` object to a standard R `list` object.
}
\section{Methods (by class)}{
\itemize{
\item \code{as.list(preproc_data)}: Method for coercion of a `preproc_data` object to a list.

}}
