% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAs.R
\name{as}
\alias{as}
\title{Convert preproc_data to preproc_data_list}
\arguments{
\item{from}{A `preproc_data` object.}
}
\value{
A `preproc_data_list` object.

A `list` object.
}
\description{
Coerces a `preproc_data` object to a `preproc_data_list` object.

Coerces a `preproc_data` object to a standard R `list` object.
}
