% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dauta4.R
\docType{data}
\name{dauta4}
\alias{dauta4}
\title{Growthrate of Algae}
\format{
Data frame with the following columns:
\describe{
  \item{species}{Phytoplankton species.}
  \item{temperature}{Temperature in °C.}
  \item{light}{Photosynthetic active radiation in
    \ifelse{html}{\out{µ E m<sup>-2</sup>s<sup>-1</sup>}}{\eqn{$\mu E\,m^{-2} s^{-1}$}{µE m^-2 s^-1}}.}
  \item{growthrate}{Population growth rate in
    \ifelse{html}{\out{d<sup>-1</sup>}}{\eqn{$d^{-1}$}{d^-1}}.}
}
}
\source{
The example data were taken manually from the figures of Dauta et al. 1990.
}
\description{
Growthrates of four different phytoplankton species, dependent on temperature
and light.
}
\references{
Dauta, A, Devaux, J, Piquemal, F and Boumnich, L (1990)
Growth Rate of Four Algae in Relation to Light and Temperature.
Hydrobiologia 207: 221–26. \doi{10.1007/BF00041459}
}
\keyword{data}
