% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_evaluate.R
\name{ml_evaluate}
\alias{ml_evaluate}
\alias{ml_evaluate,preproc_data-method}
\title{Evaluates model performance across different data subsets.}
\usage{
ml_evaluate(object, model, ...)

\S4method{ml_evaluate}{preproc_data}(
  object,
  model,
  xtype = c("both", "scale", "transform", "none"),
  ytype = c("transform", "none"),
  ...
)
}
\arguments{
\item{object}{An object of class `preproc_data` containing the data subsets.}

\item{model}{The fitted model to be evaluated.}

\item{...}{Additional arguments (currently ignored).}

\item{xtype}{Character argument if transformed input data ("transform"),
scaled data ("scale" ), transformed and scaled data ("both") or
original raw x data ("none") are used.}

\item{ytype}{Character argument if transformed target values ("transform") or
original raw y-values ("none") are used.}
}
\value{
A data frame containing the evaluation metrics for each data subset.
  Rows represent the metrics (R2, MSE, RMSE, MAE), and columns represent the
  data subsets (all, train, test).
}
\description{
This function calculates various model evaluation metrics, including R-squared,
Mean Squared Error (MSE), Root Mean Squared Error (RMSE), and Mean Absolute
Error (MAE), for different subsets of data (all, train, and test).
}
\details{
The `xtype` must be exactly the same that was used as input for the
  fitted model, while `ytype` may be "transform" (the scale used for model
  fitting) or "none", i.e. the original scale.
}
\section{Functions}{
\itemize{
\item \code{ml_evaluate(preproc_data)}: Method for estimating a set of model evaluation criteria
for a model and a `preproc_data` object.

}}
