% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R
\docType{class}
\name{feature_data-class}
\alias{feature_data-class}
\title{Preprocessed Data}
\description{
An S4 class to store the preprocessed data and its associated parameters.
}
\section{Slots}{

\describe{
\item{\code{data}}{Data frame containing the preprocessed data.}

\item{\code{params}}{`feature_params` object storing the parameters used in preprocessing.}
}}

