% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_data}
\alias{get_data}
\alias{get_data,preproc_data-method}
\title{Get All Data (X and Y)}
\usage{
get_data(object, ...)

\S4method{get_data}{preproc_data}(
  object,
  type = c("both", "scale", "transform", "none"),
  as_matrix = TRUE
)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{...}{Additional arguments (currently not used).}

\item{type}{Character argument if transformed data ("transform"),
scaled data ("scale" ), transformed and scaled data ("both") or
original raw data ("none") will be returned.}

\item{as_matrix}{logical TRUE if the function should return matrix,
or a data frame or tibble otherwise.}
}
\value{
A data frame or matrix containing all data, including `id_col` and `split_col`.
}
\description{
Access complete data set from the preprocessed data.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_data(preproc_data)}: Method for extracting all data.

}}
