% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id_train}
\alias{get_id_train}
\alias{get_id_train,preproc_data-method}
\title{Get ID Columns for training data}
\usage{
get_id_train(object, ...)

\S4method{get_id_train}{preproc_data}(object, as_data_frame = FALSE)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{...}{Additional arguments (currently not used).}

\item{as_data_frame}{logical TRUE if the function should return a data frame.}
}
\value{
A data frame or matrix containing the ID column(s).
}
\description{
Extracts ID columns (ID) from the preprocessed data.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_id_train(preproc_data)}: Method for extracting id columns for the training data

}}
