% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{get_x_train}
\alias{get_x_train}
\alias{get_x_train,preproc_data-method}
\title{Get Training Data (X)}
\usage{
get_x_train(object, ...)

\S4method{get_x_train}{preproc_data}(
  object,
  type = c("both", "scale", "transform", "none"),
  as_data_frame = FALSE
)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{...}{Additional arguments (currently not used).}

\item{as_data_frame}{logical TRUE if the function should return a data frame.}
}
\value{
A data frame or matrix containing the training data (X).
}
\description{
Extracts the training data (X) from the preprocessed data.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_x_train(preproc_data)}: Method for extracting training data (X).

}}
