% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{get_y_all}
\alias{get_y_all}
\alias{get_y_all,preproc_data-method}
\title{Get All Data (Y)}
\usage{
get_y_all(object, ...)

\S4method{get_y_all}{preproc_data}(
  object,
  type = c("transform", "none", "scale", "both"),
  as_data_frame = FALSE
)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{...}{Additional arguments (currently not used).}

\item{as_data_frame}{logical TRUE if the function should return a data frame.}
}
\value{
A data frame or matrix containing all the data (Y).
}
\description{
Extracts all the data (Y) from the preprocessed data.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_y_all(preproc_data)}: Method for extracting all data (Y).

}}
