% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_transform_data.R
\name{inv_transform_data}
\alias{inv_transform_data}
\alias{inv_transform_data,feature_data,ANY,ANY-method}
\alias{inv_transform_data,feature_data,missing,list-method}
\alias{inv_transform_data,data.frame,feature_params,missing-method}
\title{Inverse Transform Data}
\usage{
inv_transform_data(object, params, funs, ...)

\S4method{inv_transform_data}{feature_data,ANY,ANY}(object, params, funs, ...)

\S4method{inv_transform_data}{feature_data,missing,list}(object, params, funs, ...)

\S4method{inv_transform_data}{data.frame,feature_params,missing}(object, params, funs, ...)
}
\arguments{
\item{object}{A `feature_data` object.}

\item{params}{A `feature_params` object with transformation and scaling parameters}

\item{funs}{A list of inverse data transformation functions.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A `feature_data` object or a `data.frame` with the inverse transformed data.
}
\description{
Applies the inverse transformation to the data.
}
\section{Methods (by class)}{
\itemize{
\item \code{inv_transform_data(object = feature_data, params = ANY, funs = ANY)}: Method for inverse transforming data in a `feature_data` object.

\item \code{inv_transform_data(object = feature_data, params = missing, funs = list)}: Method for transforming data in a `feature_data` object and user-defined functions (deprecated).

\item \code{inv_transform_data(
  object = data.frame,
  params = feature_params,
  funs = missing
)}: Method for a `data frame`.

}}
