% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlfeatr-package.R
\docType{package}
\name{mlfeatr-package}
\alias{mlfeatr}
\alias{mlfeatr-package}
\title{Machine Learning and Feature Engineering for Regression Models}
\description{
This package provides preprocessing tools using S4 classes for data transformation, scaling, and inverse transformation.
}
\author{
\strong{Maintainer}: Thomas Petzoldt \email{thomas.petzoldt@tu-dresden.de} (\href{https://orcid.org/0000-0002-4951-6468}{ORCID})

}
