% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict,feature_data-method}
\alias{predict,feature_data-method}
\title{Model Predictions with `feature_data` Objects}
\usage{
\S4method{predict}{feature_data}(
  object,
  model,
  subset = c("all", "test", "train"),
  xprep = c("both", "scale", "transform", "none"),
  yprep = c("both", "scale", "transform", "none"),
  to_original_scale = FALSE,
  as_matrix = TRUE,
  ...
)
}
\arguments{
\item{object}{A `feature_data` object.}

\item{model}{A fitted model object.}

\item{subset}{Subset of the `preproc` data set.}

\item{xprep}{Character argument if transformed data ("transform"),
scaled data ("scale" ), transformed and scaled data ("both") or
original raw data ("none") are used for x.}

\item{yprep}{Character argument if transformed data ("transform"),
scaled data ("scale" ), transformed and scaled data ("both") or
original raw data ("none") are used for y.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A matrix or vector with the predictions.
}
\description{
Makes predictions from a machine learning model with a `feature_data` object.
}
\details{
Currently only certain Keras models can be used as fitted models.
}
