% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict,preproc_data-method}
\alias{predict,preproc_data-method}
\title{Model Predictions with `preproc_data` Objects}
\usage{
\S4method{predict}{preproc_data}(object, model, type = c("all", "test", "train"), ...)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{model}{A fitted model object.}

\item{type}{Subset of the `preproc` data set.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A matrix or vector with the predictions.
}
\description{
Makes predictions from a machine learning model with a `preproc_data` object.
}
\details{
Currently only certain Keras models can be used as fitted models.
}
