% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R
\docType{class}
\name{preproc_data_list-class}
\alias{preproc_data_list-class}
\title{List Representation of Preprocessed Data}
\description{
An S4 class to represent the preprocessed data as a list.
}
\section{Slots}{

\describe{
\item{\code{x_train}}{Training data (X).}

\item{\code{x_test}}{Test data (X).}

\item{\code{x_all}}{All data (X).}

\item{\code{y_train}}{Training data (Y).}

\item{\code{y_test}}{Test data (Y).}

\item{\code{y_all}}{All data (Y).}

\item{\code{x_train_scaled}}{Scaled training data (X).}

\item{\code{x_test_scaled}}{Scaled test data (X).}

\item{\code{x_all_scaled}}{Scaled all data (X).}
}}

