% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R
\docType{class}
\name{preproc_params-class}
\alias{preproc_params-class}
\title{Preprocessing Parameters}
\description{
An S4 class to store the parameters used in the preprocessing pipeline.
}
\section{Slots}{

\describe{
\item{\code{id_col}}{Character vector specifying the name of the ID column.}

\item{\code{target_col}}{Character vector specifying the name of the target variable column.}

\item{\code{split_col}}{Character vector specifying the name of the column used for train/test split.}

\item{\code{scale_option}}{Character string specifying the scaling option ("train", "test", "both").}

\item{\code{scale_method}}{Character string specifying the scaling method ("scale", "norm").}

\item{\code{mean_vals}}{Numeric vector storing the mean values for scaling.}

\item{\code{sd_vals}}{Numeric vector storing the standard deviation values for scaling.}

\item{\code{min_vals}}{Numeric vector storing the minimum values for normalization.}

\item{\code{max_vals}}{Numeric vector storing the maximum values for normalization.}

\item{\code{fun_transform}}{List of functions for data transformation.}

\item{\code{fun_inverse}}{List of functions for inverse data transformation.}

\item{\code{transformed}}{Boolean value, `TRUE` if object contains transformed values,
`NA` if it is the result of an inverse transformation.}
}}

