% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals,feature_data-method}
\alias{residuals,feature_data-method}
\title{Extract Model Residuals Using `feature_data` Objects}
\usage{
\S4method{residuals}{feature_data}(
  object,
  model,
  subset = c("all", "test", "train"),
  prep = c("both", "scale", "transform", "none"),
  ...
)
}
\arguments{
\item{object}{A `feature_data` object.}

\item{model}{A fitted model object.}

\item{subset}{Subset of the `preproc` data set.}

\item{prep}{Character argument if transformed data ("transform"),
scaled data ("scale" ), transformed and scaled data ("both") or
original raw data ("none") will be returned.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A matrix or vector with the residuals.
}
\description{
Extract residuals from a machine learning model compared to a `feature_data` object.
}
\details{
Currently only certain Keras models can be used as fitted models.
}
