% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals,preproc_data-method}
\alias{residuals,preproc_data-method}
\title{Extract Model Residuals Using `preproc_data` Objects}
\usage{
\S4method{residuals}{preproc_data}(object, model, type = c("all", "test", "train"), ...)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{model}{A fitted model object.}

\item{type}{Subset of the `preproc` data set.}
}
\value{
A matrix or vector with the residuals.
}
\description{
Extract residuals from a machine learning model compared to a `preproc_data` object.
}
\details{
Currently only certain Keras models can be used as fitted models.
}
