% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquared.R
\name{rsquared}
\alias{rsquared}
\alias{rsquared,preproc_data-method}
\title{Nonlinear Coefficient of Determination Using `preproc_data` Objects}
\usage{
rsquared(object, ...)

\S4method{rsquared}{preproc_data}(object, model, type = c("all", "test", "train"), ...)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{...}{Additional arguments (currently not used).}

\item{model}{A fitted model object.}

\item{type}{Subset of the `preproc` data set.}
}
\value{
The coefficient of determination.
}
\description{
Estimates the nonlinear R2 from a machine learning model compared to a
  `preproc_data` object.
}
\details{
Currently only certain Keras models can be used as fitted models.
}
\section{Methods (by class)}{
\itemize{
\item \code{rsquared(preproc_data)}: Method for estimating the coefficient of determinationof
a model for a `preproc_data` object.

}}
