% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cols.R
\name{scale_cols}
\alias{scale_cols}
\alias{scale_cols,feature_data,missing-method}
\alias{scale_cols,data.frame,feature_params-method}
\title{Scale Data}
\usage{
scale_cols(object, params, ...)

\S4method{scale_cols}{feature_data,missing}(object, params, ...)

\S4method{scale_cols}{data.frame,feature_params}(object, params, transformed = FALSE, ...)
}
\arguments{
\item{object}{A `feature_data` object or a data frame.}

\item{params}{n `feature_params` object with scaling parameters}

\item{...}{Additional arguments (currently not used).}

\item{transformed}{`TRUE` to use parameters of transformed data.}
}
\value{
A `feature_data` object with the scaled data.
}
\description{
Scales the data using the stored parameters.
}
\section{Methods (by class)}{
\itemize{
\item \code{scale_cols(object = feature_data, params = missing)}: Method for scaling data in a `feature_data` object.

\item \code{scale_cols(object = data.frame, params = feature_params)}: Method for scaling data in a `feature_data` object.

}}
