% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{scale_x}
\alias{scale_x}
\alias{scale_x,preproc_data-method}
\title{Scale Data}
\usage{
scale_x(object)

\S4method{scale_x}{preproc_data}(object)
}
\arguments{
\item{object}{A `preproc_data` object.}
}
\value{
A `preproc_data` object with the scaled data.
}
\description{
Scales the data using the stored parameters.
}
\section{Methods (by class)}{
\itemize{
\item \code{scale_x(preproc_data)}: Method for scaling data in a `preproc_data` object.

}}
