% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{transform_data}
\alias{transform_data}
\alias{transform_data,preproc_data,list-method}
\title{Transform Data}
\usage{
transform_data(object, funs)

\S4method{transform_data}{preproc_data,list}(object, funs)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{funs}{A named list of functions to apply to the data.}
}
\value{
A `preproc_data` object with the transformed data.
}
\description{
Transforms the data using the provided functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{transform_data(object = preproc_data, funs = list)}: Method for transforming data in a `preproc_data` object.

}}
