% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAs.R
\name{as}
\alias{as}
\title{Convert feature_data to feature_data_list}
\arguments{
\item{from}{A `feature_data` object.}
}
\value{
A `feature_data_list` object.

A `list` object.
}
\description{
Coerces a `feature_data` object to a `feature_data_list` object.

Coerces a `feature_data` object to a standard R `list` object.
}
