% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dauta4.R
\docType{data}
\name{dauta4}
\alias{dauta4}
\title{Growthrate of Algae}
\format{
Data frame with the following columns:
\describe{
  \item{species}{Phytoplankton species.}
  \item{temperature}{Temperature in °C.}
  \item{light}{Photosynthetic active radiation in µE $m^{-2} s^{-1}.}
  \item{growthrate}{Population growth rate in d^{-1}.}
}
}
\source{
Dauta, Alain, Jean Devaux, Françoise Piquemal, and Lhoussaine Boumnich. 1990.
Growth Rate of Four Algae in Relation to Light and Temperature. Hydrobiologia
207: 221–26. \doi{10.1007/BF00041459}
}
\description{
Growthrates of four different phytoplankton species, dependent on temperature
and light.
}
\details{
The example data were taken from the figures of Dauta et al. 1990.
}
\keyword{data}
