% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_evaluate.R
\name{ml_evaluate}
\alias{ml_evaluate}
\alias{ml_evaluate,feature_data-method}
\title{Evaluates model performance across different data subsets.}
\usage{
ml_evaluate(object, model, ...)

\S4method{ml_evaluate}{feature_data}(
  object,
  model,
  xprep = c("both", "scale", "transform", "none"),
  yprep = c("both", "scale", "transform", "none"),
  ...
)
}
\arguments{
\item{object}{An object of class `feature_data` containing the data subsets.}

\item{model}{The fitted model to be evaluated.}

\item{...}{Additional arguments (currently ignored).}

\item{xprep}{Character argument if transformed input data ("transform"),
scaled data ("scale" ), transformed and scaled data ("both") or
original raw x data ("none") are used.}

\item{yprep}{Character argument if transformed target values ("transform") or
original raw y-values ("none") are used.}
}
\value{
A data frame containing the evaluation metrics for each data subset.
  Rows represent the metrics (R2, MSE, RMSE, MAE), and columns represent the
  data subsets (all, train, test).
}
\description{
This function calculates various model evaluation metrics, including R-squared,
Mean Squared Error (MSE), Root Mean Squared Error (RMSE), and Mean Absolute
Error (MAE), for different subsets of data (all, train, and test).
}
\details{
The parameters `xprep` and `yprep` must be exactly the same
  that were used as input and output for the fitted model,
  Rescaling of `yprep` will be implemented in future versions of the package.
}
\section{Functions}{
\itemize{
\item \code{ml_evaluate(feature_data)}: Method for estimating a set of model evaluation criteria
for a model and a `feature_data` object.

}}
