% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id_train}
\alias{get_id_train}
\alias{get_id_train,feature_data-method}
\title{Get ID Columns for training data}
\usage{
get_id_train(object, ...)

\S4method{get_id_train}{feature_data}(object, as_matrix = TRUE)
}
\arguments{
\item{object}{A `feature_data` object.}

\item{...}{Additional arguments (currently not used).}

\item{as_matrix}{logical TRUE if the function should return matrix,
or a data frame or tibble otherwise.}
}
\value{
A data frame or matrix containing the ID column(s).
}
\description{
Extracts ID columns (ID) from the preprocessed data.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_id_train(feature_data)}: Method for extracting id columns for the training data

}}
