% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{get_x_test}
\alias{get_x_test}
\alias{get_x_test,preproc_data-method}
\title{Get Test Data (X)}
\usage{
get_x_test(object, ...)

\S4method{get_x_test}{preproc_data}(object, return_matrix = TRUE)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A data frame or matrix containing the test data (X).
}
\description{
Extracts the test data (X) from the preprocessed data.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_x_test(preproc_data)}: Method for extracting test data (X).

}}
