% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{get_y_test}
\alias{get_y_test}
\alias{get_y_test,preproc_data-method}
\title{Get Test Data (Y)}
\usage{
get_y_test(object, ...)

\S4method{get_y_test}{preproc_data}(
  object,
  type = c("transform", "none", "scale", "both"),
  as_matrix = TRUE
)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{...}{Additional arguments (currently not used).}

\item{as_matrix}{logical TRUE if the function should return matrix,
or a data frame or tibble otherwise.}
}
\value{
A data frame or matrix containing the test data (Y).
}
\description{
Extracts the test data (Y) from the preprocessed data.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_y_test(preproc_data)}: Method for extracting test data (Y).

}}
