% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{inverse_transform}
\alias{inverse_transform}
\alias{inverse_transform,preproc_data-method}
\title{Inverse Transform Data}
\usage{
inverse_transform(object)

\S4method{inverse_transform}{preproc_data}(object)
}
\arguments{
\item{object}{A `preproc_data` object.}
}
\value{
A `preproc_data` object with the inverse transformed data.
}
\description{
Applies the inverse transformation to the data.
}
\section{Methods (by class)}{
\itemize{
\item \code{inverse_transform(preproc_data)}: Method for inverse transforming data in a `preproc_data` object.

}}
