% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquared.R
\name{rsquared}
\alias{rsquared}
\alias{rsquared,feature_data-method}
\title{Nonlinear Coefficient of Determination Using `feature_data` Objects}
\usage{
rsquared(object, ...)

\S4method{rsquared}{feature_data}(
  object,
  model,
  subset = c("all", "test", "train"),
  prep = c("both", "scale", "transform", "none"),
  ...
)
}
\arguments{
\item{object}{A `feature_data` object.}

\item{...}{Additional arguments (currently not used).}

\item{model}{A fitted model object.}

\item{subset}{Subset of the `preproc` data set.}

\item{prep}{Character argument if transformed data ("transform"),
scaled data ("scale" ), transformed and scaled data ("both") or
original raw data ("none") will be returned.}
}
\value{
The coefficient of determination.
}
\description{
Estimates the nonlinear R2 from a machine learning model compared to a
  `feature_data` object.
}
\details{
Currently only certain Keras models can be used as fitted models.
}
\section{Methods (by class)}{
\itemize{
\item \code{rsquared(feature_data)}: Method for estimating the coefficient of determination of
a model for a `feature_data` object.

}}
