% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x.R
\name{scale_x}
\alias{scale_x}
\alias{scale_x,preproc_data-method}
\alias{scale_x,data.frame-method}
\title{Scale Data}
\usage{
scale_x(object, ...)

\S4method{scale_x}{preproc_data}(object, ...)

\S4method{scale_x}{data.frame}(object, params, ...)
}
\arguments{
\item{object}{A `preproc_data` object.}

\item{...}{Additional arguments (currently not used).}

\item{params}{An `preproc_parms` object with scaling parameters}
}
\value{
A `preproc_data` object with the scaled data.
}
\description{
Scales the data using the stored parameters.
}
\section{Methods (by class)}{
\itemize{
\item \code{scale_x(preproc_data)}: Method for scaling data in a `preproc_data` object.

\item \code{scale_x(data.frame)}: Method for scaling data in a `preproc_data` object.

}}
